refactor: move downloader interfaces to downloader dir
