+++ src/downloader/impl/jobmanager.cpp
